refactor(web): redesign the customize filter component
