+++ apps/web/src/components/issues/issue-group.tsx
@@ -34,7 +34,7 @@ export default function IssuesGroup({
   const Icon = Icons[groupData.icon as keyof typeof Icons]
 
   return (
-    <div className='flex items-center justify-between pr-2.5 my-2'>
+    <div className='flex items-center justify-between pr-2.5 mb-2'>
       <div className='flex items-center gap-2 bg-weak/50 px-3 py-1 w-32 border border-soft/50 rounded-md'>
         <Icon className='size-4 text-sub' />
         <h2 className='font-medium text-sm text-surface'>{groupData.label}</h2>
