+++ apps/web/src/components/issues/issue-list.tsx
@@ -1,10 +1,8 @@
 'use client'
 
-import { useEffect, useState } from 'react'
-
-import type { Store } from '@/lib/store/my-issues-store'
 import type { TIssue } from '@buildit/utils/types'
 
+import { useMyIssues } from '@/hooks/store'
 import { api } from '@/lib/trpc/react'
 
 import IssuesGroup from './issue-group'
@@ -12,51 +10,49 @@ import IssueItem from './issue-item'
 
 /**
  * The Issue list component to display a list of issues. And empty state for Issues Page
- * @param props The component properties.
- * @param props.store The store to use.
- * @returns React component.
+ * @returns JSX Element.
  */
-export default function IssueList({ store }: { store: Store }): JSX.Element {
+export default function IssueList(): JSX.Element {
   const { data: allIssues, error } = api.issues.get_issues.useQuery()
 
-  const [filteredIssues, setFilteredIssues] = useState<TIssue[]>([])
+  const { filters, groupBy } = useMyIssues()
 
-  const filteredStatus = store.filterByStatus
-  const filteredPriority = store.filterByPriority
-  const groupBy = store.groupBy
+  if (error) {
+    return <div>Error: {error.message}</div>
+  }
 
-  useEffect(() => {
-    const filtered = allIssues?.filter((issue) => {
-      if (filteredStatus && filteredStatus !== '') {
-        return issue.status === filteredStatus
-      }
-      if (filteredPriority && filteredPriority !== '') {
-        return issue.priority === filteredPriority
-      }
-      return true
+  const filterIssues = (issues: typeof allIssues) => {
+    if (filters.length === 0) {
+      return issues
+    }
+
+    return issues?.filter((issue) => {
+      return filters.every((filter) => {
+        switch (filter.filter) {
+          case 'status':
+            return issue.status === filter.value
+          case 'priority':
+            return issue.priority === filter.value
+          default:
+            return true
+        }
+      })
     })
+  }
 
-    if (store.search) {
-      setFilteredIssues(
-        filtered?.filter((issue) =>
-          issue.title.toLowerCase().includes(store.search.toLowerCase()),
-        ) ?? [],
-      )
-    } else {
-      setFilteredIssues(filtered ?? allIssues ?? [])
-    }
-  }, [allIssues, filteredStatus, store.search, filteredPriority])
+  const filteredIssues = filterIssues(allIssues)
 
-  const groupIssues = (issues: TIssue[], groupBy: string) => {
-    if (groupBy === 'noGrouping') {
-      return { Uncategorized: issues }
-    }
+  const groupIssues = (
+    issues: TIssue[] | undefined,
+    groupBy: keyof TIssue | 'No Grouping',
+  ) => {
+    if (!issues || groupBy === 'No Grouping') return { '': issues ?? [] }
 
-    return issues.reduce((groups: any, issue: any) => {
-      const groupKey = issue[groupBy] || 'Uncategorized'
-      if (!groups[groupKey]) {
-        groups[groupKey] = []
-      }
+    return issues.reduce((groups: Record<string, TIssue[]>, issue: TIssue) => {
+      const groupKey = (issue[groupBy] as string) || 'Uncategorized'
+
+      if (!groups[groupKey]) groups[groupKey] = []
+
       groups[groupKey].push(issue)
       return groups
     }, {})
@@ -64,36 +60,35 @@ export default function IssueList({ store }: { store: Store }): JSX.Element {
 
   const groupedIssues = groupIssues(filteredIssues, groupBy)
 
-  if (error) {
-    return <div>Error: {error.message}</div>
-  }
-
   return (
-    <div className='h-full'>
-      {allIssues?.length == 0 ? (
+    <>
+      {filteredIssues?.length === 0 ? (
         <div className='flex h-1/2 w-full flex-col items-center justify-center space-y-4 rounded-lg'>
           <div className='flex flex-col items-center'>
             <h1 className='font-cal text-strong text-xl'>No issues found</h1>
             <p className='text-sm text-sub'>
               There aren&apos;t any issues at the moment. Create one to get
-              started!{' '}
+              started!
             </p>
           </div>
         </div>
       ) : (
-        Object.keys(groupedIssues).map((group) => (
-          <div key={group}>
-            <IssuesGroup group={group} />
-            <ul className='list-none'>
-              {groupedIssues[group].map((issue: TIssue) => (
-                <li key={issue.id}>
-                  <IssueItem issue={issue} />
-                </li>
+        <>
+          {Object.entries(groupedIssues).map(([group, issues]) => (
+            <div key={group}>
+              <IssuesGroup group={group} count={issues.length} />
+              {issues.map((issue, index) => (
+                <IssueItem
+                  key={issue.id}
+                  issue={issue}
+                  isFirst={index === 0}
+                  isLast={index === issues.length - 1}
+                />
               ))}
-            </ul>
-          </div>
-        ))
+            </div>
+          ))}
+        </>
       )}
-    </div>
+    </>
   )
 }
